back/requirements.txt
flask
redis
mysql-connector
flask-cors